Extract Sinks to their own packages
Having a "Sink" interface and package facilitates the future develpoment
of new Adapters/Sinks, so that we can implement Redis/RabbitMQ or even
Kinesis/Kafka.

The abstraction helps to create unit tests mocking sink side-effects.

There is also in this commit, a new sink, for testing purposes: Void. As
the name implies it does nothing with the message, but there is an
option to print it for debugging purposes.